Fix remaining instances where VHDL integers are passed as int32_t

+++ src/jit/jit-exits.h
@@ -46,16 +46,16 @@ void x_sched_waveform_s(sig_shared_t *ss, uint32_t offset, uint64_t scalar,
                         int64_t after, int64_t reject);
 void x_file_open(int8_t *status, void **_fp, const uint8_t *name_bytes,
                  int32_t name_len, int8_t mode);
-void x_file_write(void **_fp, uint8_t *data, int32_t len);
+void x_file_write(void **_fp, uint8_t *data, int64_t len);
 int64_t x_file_read(void **_fp, uint8_t *data, int64_t size, int64_t count);
 void x_file_close(void **_fp);
 int8_t x_endfile(void *_f);
-void x_index_fail(int32_t value, int32_t left, int32_t right, int8_t dir,
+void x_index_fail(int64_t value, int64_t left, int64_t right, int8_t dir,
                   tree_t where, tree_t hint);
-void x_length_fail(int32_t left, int32_t right, int32_t dim, tree_t where);
+void x_length_fail(int64_t left, int64_t right, int32_t dim, tree_t where);
 void x_range_fail(int64_t value, int64_t left, int64_t right, int8_t dir,
                   tree_t where, tree_t hint);
-void x_exponent_fail(int32_t value, tree_t where);
+void x_exponent_fail(int64_t value, tree_t where);
 void x_overflow(int64_t lhs, int64_t rhs, tree_t where);
 void x_null_deref(tree_t where);
 void x_div_zero(tree_t where);
@@ -89,7 +89,7 @@ void x_disconnect(sig_shared_t *ss, uint32_t offset, int32_t count,
 void x_force(sig_shared_t *ss, uint32_t offset, int32_t count, void *values);
 void x_release(sig_shared_t *ss, uint32_t offset, int32_t count);
 void x_resolve_signal(sig_shared_t *ss, jit_handle_t handle, void *context,
-                      int32_t ileft, int32_t nlits, int32_t flags);
+                      int64_t ileft, int32_t nlits, int32_t flags);
 void x_elab_order_fail(tree_t where);
 void x_unreachable(tree_t where);
 void x_cover_setup_toggle_cb(sig_shared_t *ss, int32_t *toggle_mask);
